+++ src/app/models/recording.ts
@@ -230,8 +230,8 @@ export class Recording {
       {
         phone_number: groups['phone_number'],
         phone_number_formatted: groups['phone_number'],
-        caller_name: groups['caller_name'] ? groups['caller_name'] : groups['phone_number'],
-        contact_name: groups['caller_name'] ? groups['caller_name'] : groups['phone_number'],
+        caller_name: groups['caller_name'] || groups['phone_number'],
+        contact_name: groups['caller_name'] || groups['phone_number'],
       }
     ];